refacrtor move testHelper to test folder
